Move some methods out of the Interceptor
